+++ chicken-services/chicken-eligibility/chickenEligibility.js
@@ -7,12 +7,13 @@ module.exports = function(chicken_table, acc_nber, client_json){
     
     if(chickenCursor.hasNext()){
         var row = chickenCursor.next();
-        prepRequired = row.vars.prep_required;
+        //prepRequired = row.vars.prep_required; //Replaced by percentage healthpath from november
         state.vars.chcken_nber = row.vars.ordered_chickens || 0;
         state.vars.farmer_name  = JSON.parse(state.vars.client_json).FirstName;
     }
     else{
         state.vars.client_notfound = true;
+        return;
         //var logger = new Log();
         //var logMessage = 'Client ' + client_json.FirstName + ' not found in chicken table';
         //logger.log(logMessage);
@@ -21,6 +22,9 @@ module.exports = function(chicken_table, acc_nber, client_json){
     if(client_json.BalanceHistory.length>0){
         client_json.BalanceHistory = client_json.BalanceHistory[0];
     }
+    //new prepayment calculation from November
+    prepRequired =  (client_json.BalanceHistory.TotalCredit * parseInt(project.vars.chickenRequiredPercentage))/100;
+    console.log('prep required:'+ prepRequired);
     var prepayment_amount = client_json.BalanceHistory.TotalRepayment_IncludingOverpayments - prepRequired;
     // if prepayment satisfies the mminimum condition( > than 2 chicken prepayment amount(2000))
     if(prepayment_amount >= 1000){